% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-footer.R
\name{add_footnotes}
\alias{add_footnotes}
\title{Add \code{hf_line} footnote(s) to a \code{rtf_doc} object}
\usage{
add_footnotes(doc, ...)
}
\arguments{
\item{doc}{\code{rtf_doc} object to add header/footer lines to}

\item{...}{A vector of \code{hf_line} objects to add.}
}
\value{
\code{rtf_doc} object with \code{hf_line} objects attached.
}
\description{
Add \code{hf_line} footnote(s) to a \code{rtf_doc} object
}
\examples{
# Adding lines after rtf_doc construction
ht <- huxtable::huxtable(
 column1 = 1:5,
 column2 = letters[1:5]
)
rtf <- rtf_doc(ht)

rtf <- add_footnotes(rtf, hf_line("The Footnote"))
}
\seealso{
\code{\link{add_hf}}
}
