% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{align}
\alias{align}
\alias{align<-}
\alias{set_align}
\title{Return and set text alignment}
\usage{
align(x, ...)

align(x) <- value

set_align(x, value)
}
\arguments{
\item{x}{\code{hf_line} object}

\item{...}{Additional arguments passed to method dispatch.}

\item{value}{A string representing the alignment.}
}
\value{
Alignment of the supplied \code{hf_line} object.
}
\description{
These functions modify and return the text alignment in a \code{hf_line}
object. Supported options are: 'left', 'right', 'center', and 'split'.
}
