% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{font}
\alias{font}
\alias{font<-}
\alias{set_font}
\title{Return or set font}
\usage{
font(x, ...)

font(x) <- value

set_font(x, value)
}
\arguments{
\item{x}{\code{rtf_doc} object, the table of a \code{rtf_doc} object, or a
\code{hf_line} object}

\item{...}{Additional arguments passed to method dispatch}

\item{value}{A string representing a font}
}
\value{
For \code{font()}, the font attribute of the object in the case of
  \code{hf_line} and the table, or each unique font in the table, titles,
  footnotes, and the overall document in the case of \code{rtf_doc}. For
  \code{set_font()} and \code{`font<-`()}, the modified object.
}
\description{
These functions modify and return the fonts of a \code{rtf_doc} object and
associated items. To set the font attribute of a table you should use the
functions assosiated with that package and modify the table directly.
}
