% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_text_string}
\alias{format_text_string}
\title{Title}
\usage{
format_text_string(text, properties = "")
}
\arguments{
\item{text}{text}

\item{properties}{properties}
}
\value{
string
}
\description{
Title
}
