% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_component_extractors.R
\name{get_table_body}
\alias{get_table_body}
\title{Return the RTF string for the body of a table}
\usage{
get_table_body(doc)
}
\arguments{
\item{doc}{\code{rtf_doc} object to return header string from.}
}
\value{
A RTF encoding string containing the header information
}
\description{
This function relys on the package that the table originated from to create
a RTF string.
}
