% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-footer.R
\name{new_hf_line}
\alias{new_hf_line}
\title{Create a title line container}
\usage{
new_hf_line(line, align, bold, italic, font, font_size, index)
}
\arguments{
\item{line}{A character list/vector}

\item{align}{Alignment in the document}

\item{bold}{Is bold?}

\item{italic}{Is italic?}

\item{font}{Font used in document, character}

\item{font_size}{pt of font in document, numeric}

\item{index}{order in document}
}
\value{
An object of class hf_line
}
\description{
Create a title line container
}
