% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_doc.R
\name{new_rtf_doc}
\alias{new_rtf_doc}
\title{Create an Rich Text Format table document}
\usage{
new_rtf_doc(table, titles, footnotes)
}
\arguments{
\item{table}{A table of a supported class}

\item{titles}{An object/list of /code{hf_line}}

\item{footnotes}{An object/list of /code{hf_line}}
}
\value{
A list with a table, titels, and footnotes component. Class of "rtf_doc"
}
\description{
Create an Rich Text Format table document
}
\seealso{
Other rtf_doc: 
\code{\link{as_rtf_doc}()},
\code{\link{validate_rtf_doc}()}
}
\concept{rtf_doc}
