% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-footer.R
\name{order_lines}
\alias{order_lines}
\title{Order header/footer lines in an rtf_document}
\usage{
order_lines(lines)
}
\arguments{
\item{lines}{A list/vector of \code{hf_line} objects}
}
\value{
Reordered lines based on \code{index}
}
\description{
Order header/footer lines in an rtf_document
}
