% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{replace_cell_padding}
\alias{replace_cell_padding}
\title{Create an Rich Text Format table document}
\usage{
replace_cell_padding(txt)
}
\arguments{
\item{txt}{text of the rtf document}
}
\value{
replaced text without padding flags
}
\description{
Create an Rich Text Format table document
}
