% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-footer.R, R/read_hf.R
\name{titles_and_footnotes_from_df}
\alias{titles_and_footnotes_from_df}
\title{Read titles and footnotes from a dataframe}
\usage{
titles_and_footnotes_from_df(
  doc,
  from.df = NULL,
  from.file = NULL,
  reader = NULL,
  ...
)

titles_and_footnotes_from_df(
  doc,
  from.df = NULL,
  from.file = NULL,
  reader = NULL,
  ...
)
}
\arguments{
\item{doc}{A \code{rtf_doc} object to append header and footnote information.}

\item{from.df}{\code{data.frame} containing information to convert to
\code{hf_line} object(s).}

\item{from.file}{A file that contains information to convert to
\code{hf_line} object(s). Information is extracted from the file with the
supplied reader.}

\item{reader}{A function to import the information from a file supplied.}

\item{...}{Parameters passed to \code{read_hf} where they are processed and
constructed into \code{hf_line} objects.}
}
\value{
A \code{rtf_doc} object with header/footnote information attached.

A \code{rtf_doc} object with header/footnote information attached.
}
\description{
Reads a data frame with header/footnote information and attaches it to an
  \code{rtf_doc} object.The most effective way to use this function is to
  pass information to a custom reader for your process. See <Vignette Link
  here>

Reads a data frame with header/footnote information and attaches it to an
  \code{rtf_doc} object.The most effective way to use this function is to pass information to a
  custom reader for your process. See <Vignette Link here>
}
\section{Required Columns}{

\itemize{
\item{type}
\item{text1}
\item{text2}
\item{align}
\item{bold}
\item{italic}
\item{font}
\item{index}
}


\itemize{
\item{type}
\item{text1}
\item{text2}
\item{align}
\item{bold}
\item{italic}
\item{font}
\item{index}
}
}

\seealso{
\code{\link{read_hf}} reads in each line.

[read_hf()] reads in each line.
}
