% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hf.R
\name{validate_hf_dataframe}
\alias{validate_hf_dataframe}
\title{Title}
\usage{
validate_hf_dataframe(df, required_columns)
}
\arguments{
\item{df}{df}

\item{required_columns}{required_columns}
}
\value{
hf
}
\description{
Title
}
