% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-footer.R
\name{validate_hf_line}
\alias{validate_hf_line}
\title{Validate a new title line container}
\usage{
validate_hf_line(line, align, bold, italic, font, font_size, index)
}
\arguments{
\item{align}{Alignment in the document}

\item{bold}{Is bold?}

\item{italic}{Is italic?}

\item{font}{Font used in document, character}

\item{font_size}{pt of font in document, numeric}

\item{index}{order in document}

\item{...}{A character list/vector}
}
\description{
Validate a new title line container
}
