% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_doc.R
\name{validate_rtf_doc}
\alias{validate_rtf_doc}
\title{Validate parameters passed to rtf_doc}
\usage{
validate_rtf_doc(table, titles, footnotes)
}
\arguments{
\item{titles}{An object/list of /code{hf_line}}

\item{footnotes}{An object/list of /code{hf_line}}

\item{tables}{A table of a supported class}

\item{header.rows}{An integer determining how many rows of the table are headers.}
}
\value{
nothing for now
}
\description{
Validate parameters passed to rtf_doc
}
\seealso{
Other rtf_doc: 
\code{\link{as_rtf_doc}()},
\code{\link{new_rtf_doc}()}
}
\concept{rtf_doc}
