% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{view_footnotes}
\alias{view_footnotes}
\title{View footnote information}
\usage{
view_footnotes(doc)
}
\arguments{
\item{doc}{\code{rtf_doc} object}
}
\value{
\code{data.frame} of the footnote information
}
\description{
View footnote information
}
