% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{view_hf}
\alias{view_hf}
\title{View header and footer information}
\usage{
view_hf(doc, type = NULL)
}
\arguments{
\item{doc}{RTF document}

\item{type}{type}
}
\description{
View header and footer information
}
