% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{view_titles}
\alias{view_titles}
\title{View title information}
\usage{
view_titles(doc)
}
\arguments{
\item{doc}{\code{rtf_doc} object}
}
\value{
\code{data.frame} of the title information
}
\description{
View titles attached to an \code{rtf_doc} as a data.frame.
}
