% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-footer.R
\name{add_footnotes}
\alias{add_footnotes}
\title{Add \code{hf_line} object(s) to a \code{rtf_doc} object}
\usage{
add_footnotes(doc, ...)
}
\arguments{
\item{doc}{\code{rtf_doc} object to add header/footer lines to}

\item{...}{A vector of \code{hf_line} objects to add.}
}
\value{
\code{rtf_doc} object with \code{hf_line} objects attached.
}
\description{
Add \code{hf_line} object(s) to a \code{rtf_doc} object
}
\seealso{
Other hf_line: 
\code{\link{add_hf}()},
\code{\link{add_titles}()},
\code{\link{hf_line}()}
}
\concept{hf_line}
