% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{footer_height}
\alias{footer_height}
\alias{footer_height<-}
\alias{set_footer_height}
\title{Return and set footer height}
\usage{
footer_height(x, ...)

footer_height(x) <- value

set_footer_height(x, value)
}
\arguments{
\item{x}{A \code{rtf_doc} object}

\item{...}{Additonal arguments passed to method dispatch}

\item{value}{A numeric value to set the footer_height}
}
\value{
For \code{footer_height}, the footer_height attribute of the
  supplied \code{rtf_doc} object. For \code{`footer_height<-`()} and
  \code{set_footer_height()}, the modified object.
}
\description{
These functions modify and return the footer_height attribute of a \code{rtf_doc}
object.
}
