% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{header_rows}
\alias{header_rows}
\alias{header_rows<-}
\alias{set_header_rows}
\title{Return and set the header.rows}
\usage{
header_rows(x, ...)

header_rows(x) <- value

set_header_rows(x, value)
}
\arguments{
\item{x}{\code{rtf_doc} object}

\item{...}{Additonal arguments passed to method dispatch}

\item{value}{A numeric value to change the header.rows attribute.}
}
\value{
return
}
\description{
These functions modify and return the header.rows attribute of a \code{rtf_doc}
object.
}
