% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-footer.R
\name{hf_line}
\alias{hf_line}
\title{Create a title line container}
\usage{
hf_line(
  ...,
  align = c("center", "left", "right", "split"),
  bold = FALSE,
  italic = FALSE,
  font = NA,
  font_size = NaN,
  index = NULL
)
}
\arguments{
\item{...}{A character list/vector. If \code{length(...)} is 2 and
\code{align} is not 'split', values are pasted together.}

\item{align}{Text alignment as left, right, center, or split. Defaults to center.}

\item{bold}{\code{TRUE} or  \code{FALSE}. Defaults to FALSE.}

\item{italic}{\code{TRUE} or  \code{FALSE}. Defaults to FALSE.}

\item{font}{A string to specify the font display. Ensure the intended RTF
reader can display the selected font.}

\item{font_size}{Font size in half points. For example font_size = 20
will display a 10 point font. Defaults to a 12 point font.}

\item{index}{Position to display header or footnote lines in the RTF
document. Orderes in ascending order with NULLs last.}
}
\value{
An object of class \code{hf_line}
}
\description{
\code{hf_line} objects are passed to \code{rtf_doc} for display in the RTF
document. A character vector of length <= 2 describes the text to display.
}
\section{Supported Formatting}{

Several page formats are supported to display document data. When the
\code{rtf_doc} is written, the package will determine if the text of an
\code{hf_line} object starts with a keyword. Regular expresion matching and
replacement is used for formatting.
\itemize{
\item{PAGE_FORMAT: - Can take up to two replacements to format current
  page(first), and total number of pages(second)}
\item{DATE_FORMAT: - Describes the date/time the document was generated}
\item{FILE_PATH: - Describes the file path the R session was executed from}
}
}

\examples{
# Adding lines during rtf_doc construction
ht <- huxtable::huxtable(
 column1 = 1:5,
 column2 = letters[1:5]
)
rtf <- rtf_doc(ht)
titles_l <- list(
hf_line("The Title"),
# the below will display 'Current Page 1 Total Pages 1
hf_line("PAGE_FORMAT: Current Page \%s Total Pages \%s")
)
rtf <- rtf_doc(ht, titles = titles_l)

# Adding lines after rtf_doc construction
rtf <- add_footnotes(rtf, hf_line("The Footnote"))

}
