% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-footer.R
\name{hf_line}
\alias{hf_line}
\title{Create a title line container}
\usage{
hf_line(
  ...,
  align = c("center", "left", "right", "split"),
  bold = FALSE,
  italic = FALSE,
  font = NA,
  font_size = NaN,
  index = NULL
)
}
\arguments{
\item{...}{A character list/vector. If \code{length(...)} is 2 and
\code{align} is not 'split', values are pasted together.}

\item{align}{Alignment in the document. Supports 'left', 'right', 'center'
and 'split'}

\item{bold}{Bold - \code{TRUE} or \code{FALSE}}

\item{italic}{Italic - \code{TRUE} or \code{FALSE}}

\item{font}{Font used in document, string}

\item{font_size}{pt of font in document, numeric}

\item{index}{order in document, numeric.}
}
\value{
An object of class \code{hf_line}
}
\description{
\code{hf_line} objects are passed to \code{rtf_doc} for display in the RTF
document. A list of 0, 1, or 2 strings with attributes for display.
}
\details{
Currently supports the following properties:
\itemize{
\item {Alignment}
\item {Bold}
\item {Italic}
\item {Font}
\item {Font Size}
\item {Index}
}
}
\examples{
# Adding lines during rtf_doc construction
ht <- huxtable(
 column1 = 1:5,
 column2 = letters[1:5]
)
rtf <- rtf_doc(ht)
titles_l <- list(
hf_line("The Title")
)
rtf <- rtf_doc(ht, titles = titles_l)

# Adding lines after rtf_doc construction
rtf <- add_footnotes(rtf, hf_line("The Footnote"))

}
\seealso{
Other hf_line: 
\code{\link{add_footnotes}()},
\code{\link{add_hf}()},
\code{\link{add_titles}()}
}
\concept{hf_line}
