% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-footer.R
\name{hf_line}
\alias{hf_line}
\title{Create a title line container}
\usage{
hf_line(
  ...,
  align = c("center", "left", "right", "split"),
  bold = FALSE,
  italic = FALSE,
  font = NA,
  font_size = 12,
  index = NULL
)
}
\arguments{
\item{...}{A character list/vector}

\item{align}{Alignment in the document}

\item{bold}{Is bold?}

\item{italic}{Is italic?}

\item{font}{Font used in document, character}

\item{font_size}{pt of font in document, numeric}

\item{index}{order in document}
}
\value{
An object of class hf_line
}
\description{
Create a title line container
}
