% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{index}
\alias{index}
\alias{index<-}
\alias{set_index}
\title{Return or set index}
\usage{
index(x, ...)

index(x) <- value

set_index(x, value)
}
\arguments{
\item{x}{\code{hf_line} object}

\item{...}{Additonal arguments passed to method dispatch}

\item{value}{Numeric value to order index}
}
\value{
return
}
\description{
These functions modify and return the index of a \code{hf_line} object.
}
