% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{pagesize}
\alias{pagesize}
\alias{pagesize<-}
\alias{set_pagesize}
\title{Return and set pagesize}
\usage{
pagesize(x, ...)

pagesize(x) <- value

set_pagesize(x, value)
}
\arguments{
\item{x}{A \code{rtf_doc} object}

\item{...}{Additonal arguments passed to method dispatch}

\item{value}{A named numeric vector with the names height and width.}
}
\value{
For \code{pagesize()}, the pagesize attribute of the \code{rtf_doc}
  object. For \code{`pagesize<-`()} and \code{set_pagesize()}, the modified
  object.
}
\description{
These functions modify and return the pagesize attribute of a \code{rtf_doc}
object. Stored as a named vector with height and width names.
}
