% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmaRTF.R
\docType{package}
\name{pharmaRTF-package}
\alias{pharmaRTF}
\alias{pharmaRTF-package}
\title{pharmaRTF}
\description{
Enhanced RTF wrapper written in R for use with existing R tables packages
such as huxtable or GT. This package fills a gap where certain packages could
be written out as RTF, but couldn't add certain metadata or features to the
document that are required/expected in a report for a regulatory submission

This package intends to provide a flexible and reliable framework to connect
R to a pharmaceutical reporting workflow.
}
\section{Future Plans}{

\itemize{
\item{colors}
\item{graphs}
}
}

\author{
\strong{Maintainer}: Michael Stackhouse \email{mike.stackhouse@atorusresearch.com}

Authors:
\itemize{
  \item Eli Miller \email{eli.miller@atorusresearch.com} (\href{https://orcid.org/0000-0002-2127-9456}{ORCID})
  \item Ashley Tarasiewicz \email{Ashley.Tarasiewicz@atorusresearch.com}
}

}
