% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hf.R
\name{read_hf}
\alias{read_hf}
\title{Title}
\usage{
read_hf(from = NULL, from.file = NULL, type = NULL, reader = NULL, ...)
}
\arguments{
\item{from}{from}

\item{from.file}{from.file}

\item{type}{type}

\item{reader}{reader}

\item{...}{...}
}
\value{
hf
}
\description{
Title
}
