% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_doc.R
\name{rtf_doc}
\alias{rtf_doc}
\title{Create an Rich Text Format table document}
\usage{
rtf_doc(table, titles = list(), footnotes = list(), header.rows = 1)
}
\arguments{
\item{table}{A table of a supported class}

\item{titles}{An object/list of \code{hf_line}}

\item{footnotes}{An object/list of \code{hf_line}}

\item{header.rows}{An integer determining how many rows of the table are headers.}
}
\value{
A list with a table, titles, and footnotes component. Class of "rtf_doc"
}
\description{
Create an Rich Text Format table document
}
