% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_doc.R
\name{rtf_doc}
\alias{rtf_doc}
\title{Create an \code{rtf_doc} object}
\usage{
rtf_doc(table, titles = list(), footnotes = list(), header.rows = 1)
}
\arguments{
\item{table}{A table of a supported class}

\item{titles}{A list of \code{hf_line} objects.}

\item{footnotes}{An object/list of \code{hf_line}}

\item{header.rows}{An integer determining how many rows of the table are headers.}
}
\value{
A list with a table, titles, and footnotes component. Class of "rtf_doc"
}
\description{
This constructs the main object that will be used to write the RTF document.
The object is composed of a table, titles(s), and footnote(s).

A table should be a supported class. The huxtable package is the most supported,
however there are methods for the GT package. The \code{rtf_doc} will inherit
the properties from the table passed.

The titles and footnotes are composed of \code{hf_line} objects.
}
\section{\code{rtf_doc} Properties}{

Document level properties will be used where they are not overriden by
\code{hf_line} or table properties.
\itemize{
\item{font}
\item{font Size}
\item{margins}
\item{orientation}
\item{header_height}
\item{footer_height}
\item{pagesize}
\item{header.rows}
\item{ignore_cell_padding}
}
}

\examples{
# Adding lines during rtf_doc construction
ht <- huxtable::huxtable(
 column1 = 1:5,
 column2 = letters[1:5]
)
rtf <- rtf_doc(ht, titles = list(hf_line("My Header")))

}
\seealso{
[hf_line()]
}
