% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-footer.R
\name{titles_and_footnotes_from_df}
\alias{titles_and_footnotes_from_df}
\title{Read titles and footnotes from a dataframe}
\usage{
titles_and_footnotes_from_df(doc, ...)
}
\arguments{
\item{doc}{RTF document}

\item{...}{header and footer inormation}
}
\description{
Read titles and footnotes from a dataframe
}
