% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-footer.R
\name{titles_and_footnotes_from_df}
\alias{titles_and_footnotes_from_df}
\title{Read titles and footnotes from a dataframe}
\usage{
titles_and_footnotes_from_df(doc, ...)
}
\arguments{
\item{doc}{A \code{rtf_doc} object to append header and footnote information.}

\item{...}{Parameters passed to \code{read_hf} where they are processed and
constructed into \code{hf_line} objects.}
}
\value{
A \code{rtf_doc} object with header/footnote information attached.
}
\description{
Reads a data frame with header/footnote information and attaches it to an
  \code{rtf_doc} object.The most effective way to use this function is to pass information to a
  custom reader for your process. See <Vignette Link here>
}
\section{Required Columns}{

\itemize{
\item{type}
\item{text1}
\item{text2}
\item{align}
\item{bold}
\item{italic}
\item{font}
\item{index}
}
}

\seealso{
[read_hf()] reads in each line.
}
